modules/graph/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "karrio_server_graph"
version = "2025.5rc29"
description = "Multi-carrier shipping API Graph module"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    {name = "karrio", email = "hello@karrio.io"}
]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "karrio_server_core",
    "django-filter",
    "strawberry-graphql",
]

[project.urls]
Homepage = "https://github.com/karrioapi/karrio"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
exclude = ["tests.*", "tests"]
namespaces = true

[tool.setuptools.package-data]
"karrio.server.graph" = ["templates/**/*"]
"*" = ["*.html", "*.txt", "*.css", "*.js", "*.png", "*.jpg", "*.gif", "*.ico"]
